Move RTTR_Assert and some defines into commonSrc
