+++ driver/video/GLFW/src/CMakeLists.txt
@@ -5,9 +5,8 @@ IF (${CMAKE_SYSTEM_NAME} STREQUAL "Linux" AND ${PLATFORM_ARCH} STREQUAL "i386")
 		FIND_PACKAGE(Boost 1.55.0 REQUIRED)
 		INCLUDE_DIRECTORIES(${Boost_INCLUDE_DIR})
 
-		ADD_LIBRARY(videoGLFW SHARED GLFW.cpp GLFW.h ${PROJECT_SOURCE_DIR}/driver/src/DriverInterface.cpp)
-		ADD_DEPENDENCIES(videoGLFW updateversion)	
-		TARGET_LINK_LIBRARIES(videoGLFW videodrv ${GLFW_LIBRARY})
+		ADD_LIBRARY(videoGLFW SHARED GLFW.cpp GLFW.h ${PROJECT_SOURCE_DIR}/driver/src/DriverInterface.cpp ${COMMON_SRC})
+		TARGET_LINK_LIBRARIES(videoGLFW videodrv s25util ${GLFW_LIBRARY})
 	
 		SET(_OUTPUT_DIR "${CMAKE_BINARY_DIR}/${RTTR_DRIVERDIR}/video")
 		SET_TARGET_PROPERTIES(videoGLFW PROPERTIES
