+++ src/macros.h
@@ -20,33 +20,6 @@
 #ifndef MACROS_H_INCLUDED
 #define MACROS_H_INCLUDED
 
-// Macro that can be used to suppress unused warnings. Required e.g. for const boost::arrays defined in headers
-// Don't use this if not absolutely necessary!
-#ifdef __GNUC__
-#define SUPPRESS_UNUSED __attribute__((unused))
-#else
-#define SUPPRESS_UNUSED
-#endif
-
-// RTTR_FUNCTION_NAME evaluates to the name of the current function if supported
-#if defined(__STDC_VERSION__) && (__STDC_VERSION__ >= 199901L) /* C99 supports __func__ as a standard. */
-#define RTTR_FUNCTION_NAME __func__
-#elif((__GNUC__ >= 2) || defined(_MSC_VER))
-#define RTTR_FUNCTION_NAME __FUNCTION__
-#else
-#define RTTR_FUNCTION_NAME "<Unknown Func>"
-#endif
-
-#if defined _WIN32 && defined _DEBUG && defined _MSC_VER && !defined NOCRTDBG
-// Check for heap corruption
-#define CHECK_HEAP_CORRUPTION _ASSERTE(_CrtCheckMemory());
-#else
-#define CHECK_HEAP_CORRUPTION
-#endif // _WIN32 && _DEBUG && !NOCRTDBG
-
-/// Call a member function trough an object and a member function pointer
-#define CALL_MEMBER_FN(object, ptrToMember) ((object).*(ptrToMember))
-
 /// Iterate over all points of an area using a point of TYPE named "pt"
 /// WIDTH and HEIGHT is evaluated at most once
 /// Use like:
