+++ src/test/testBuilding.cpp
@@ -19,8 +19,8 @@
 #include "GameClient.h"
 #include "GameMessages.h"
 #include "GamePlayer.h"
-#include "RTTR_AssertError.h"
 #include "buildings/nobBaseMilitary.h"
+#include "commonSrc/RTTR_AssertError.h"
 #include "desktops/dskGameInterface.h"
 #include "initTestHelpers.h"
 #include "world/GameWorldViewer.h"
