+++ src/test/testSeaWorldCreation.cpp
@@ -16,8 +16,8 @@
 // along with Return To The Roots. If not, see <http://www.gnu.org/licenses/>.
 
 #include "defines.h" // IWYU pragma: keep
-#include "RTTR_AssertError.h"
 #include "SeaWorldWithGCExecution.h"
+#include "commonSrc/RTTR_AssertError.h"
 #include <boost/test/unit_test.hpp>
 
 std::ostream& operator<<(std::ostream& out, const ShipDirection& dir)
